Profiles refactor (#403)
This PR Refactors models, views, and serializers into module folders. It is basically done, but I bet there's room for improvement and places where imports are inconsistent.
