+++ pulseapi/profiles/models/profiles.py
@@ -1,9 +1,11 @@
 import os
 
-from django.db import models
 from django.conf import settings
-from django.utils import timezone
+from django.db import models
 from django.utils.html import format_html
+from django.utils import timezone
+
+from . import ProfileType
 
 
 def entry_thumbnail_path(instance, filename):
@@ -11,66 +13,6 @@ def entry_thumbnail_path(instance, filename):
         timestamp=str(timezone.now()),
         ext=os.path.splitext(filename)[1]
     )
-
-
-class ProfileType(models.Model):
-    """
-    See https://github.com/mozilla/network-pulse/issues/657
-
-    Values that should exist (handled via migration):
-
-    - plain
-    - staff
-    - fellow
-    - board member
-    - grantee
-    """
-    value = models.CharField(
-        max_length=50,
-        unique=True
-    )
-
-    def get_default_profile_type():
-        (default, _) = ProfileType.objects.get_or_create(value='plain')
-        return default
-
-    def __str__(self):
-        return self.value
-
-
-class ProgramType(models.Model):
-    """
-    See https://github.com/mozilla/network-pulse/issues/657
-
-    These values are determined by pulse API administrators
-    (tech policy fellowship, mozfest speaker, etc)
-    """
-    value = models.CharField(
-        max_length=150,
-        unique=True
-    )
-
-    def __str__(self):
-        return self.value
-
-
-class ProgramYear(models.Model):
-    """
-    See https://github.com/mozilla/network-pulse/issues/657
-
-    You'd think this would be 4 characters, but a "year" is
-    not a calendar year, so the year could just as easily
-    be "summer 2017" or "Q2 2016 - Q1 2018", so this is the
-    same kind of simple value model that the profile and
-    program types use.
-    """
-    value = models.CharField(
-        max_length=25,
-        unique=True
-    )
-
-    def __str__(self):
-        return self.value
 
 
 class UserProfileQuerySet(models.query.QuerySet):
@@ -295,37 +237,3 @@ class UserProfile(models.Model):
 
     class Meta:
         verbose_name = "Profile"
-
-
-class UserBookmarks(models.Model):
-    """
-    This class is used to link users and entries through a
-    "bookmark" relation. One user can bookmark many entries,
-    and one entry can have bookmarks from many users.
-    """
-    entry = models.ForeignKey(
-        'entries.Entry',
-        on_delete=models.CASCADE,
-        related_name='bookmarked_by'
-    )
-
-    profile = models.ForeignKey(
-        'profiles.UserProfile',
-        on_delete=models.CASCADE,
-        related_name='bookmarks_from',
-        null=True
-    )
-
-    timestamp = models.DateTimeField(
-        auto_now=True,
-    )
-
-    def __str__(self):
-        return 'bookmark for "{e}" by [{p}]'.format(
-            e=self.entry,
-            p=self.profile.id
-        )
-
-    class Meta:
-        verbose_name = "Bookmarks"
-        verbose_name_plural = "Bookmarks"
