+++ pulseapi/profiles/serializers/bookmarks.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+
+from pulseapi.profiles.models import UserBookmarks
+
+
+class UserBookmarksSerializer(serializers.ModelSerializer):
+    """
+    Serializes a {user,entry,when} bookmark.
+    """
+
+    class Meta:
+        """
+        Meta class. Again: because
+        """
+        model = UserBookmarks
