+++ pulseapi/profiles/views/__init__.py
@@ -1,0 +1,17 @@
+from pulseapi.profiles.views.profiles import (
+    UserProfileAPIView,
+    UserProfilePublicAPIView,
+    UserProfilePublicSelfAPIView,
+    UserProfileListAPIView,
+)
+from pulseapi.profiles.views.categories import UserProfileCategoriesView
+from pulseapi.profiles.views.profile_entries import UserProfileEntriesAPIView
+
+__all__ = [
+    'UserProfileAPIView',
+    'UserProfileCategoriesView',
+    'UserProfileEntriesAPIView',
+    'UserProfileListAPIView',
+    'UserProfilePublicAPIView',
+    'UserProfilePublicSelfAPIView'
+]
